multi-monitor support
Added support for multiple monitors. While sxiv defaults to full-screen, this is limited to the current monitor. Adding `-Z` ensures that if you're on the top-left monitor, this works as expected. Monitors will open full-screen views of the screenshot with the top left pixel of the _monitor_ showing the top left pixel of the screenshot, which is only appropriate on the top-left-most monitor.

I also used `trap` to ensure `$tmpfile` is cleaned up regardless of whether the script is terminated. 

Nice observations on speed! `maim` is a _lot_ faster than ImageMagick's `import` and `sxiv` is quite speedy as well. The script will be slightly faster using `/bin/sh` over `/bin/bash` (which would be slightly faster than `/usr/bin/env bash`), and since there are no bashisms, the more basic POSIX shell is functionally identical.